Don't require scalable font for secondary fonts
I found that requiring FC_SCALABLE would mean preferring other font
faces/families when loading secondary fonts, even if the requested font
pattern does have fonts that would satisfy the character. This is the
case in my testing with Terminus at least. We load iso-8859-1 version of
terminus for some reason, but if we require a Unicode code point that
isn't in that, we can load the Unicode version of Terminus with this
change. Before this change, we would fall back to another font.

I also add setlocale(LC_CTYPE, "") though this doesn't change behaviour.
I think it is just a good practice (it does change behaviour if using
Xlib fonts).